+++ Frontend/src/lib/images/index.ts
@@ -10,5 +10,5 @@ export {default as info} from './info.svg';
 export {default as alert} from './alert.svg';
 export {default as success} from './success.svg';
 
-export {default as circuitBoard} from './circuit-board.svg'
-export {default as itCrowd} from './IT-Crowd.gif'
+export {default as circuitBoard} from './circuit-board.svg';
+export {default as itCrowd} from './IT-Crowd.gif';